requirements.txt
# Core dependencies
solana>=0.31.0
spl-token>=0.10.0
python-dotenv>=1.0.0
pywebview>=4.0
e2b_desktop>=0.5.0

# API and data processing
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0

# Automation and web interaction
smolagents>=0.1.0
selenium>=4.10.0
helium>=3.2.8
pillow>=10.0.0

# Security and encryption
cryptography>=41.0.0

# Testing and monitoring
pytest>=7.3.1
pytest-cov>=4.1.0

# Utilities
tqdm>=4.65.0
retrying>=1.3.4
